fix: route MCP catalog lookup through DB management
